# internal/api/handlers/comment_handler.go
package handlers

import (
	"net/http"

	"github.com/davidrdsilva/blog-api/internal/application/dtos"
	"github.com/davidrdsilva/blog-api/internal/application/services"
	"github.com/davidrdsilva/blog-api/internal/domain/models"
	"github.com/davidrdsilva/blog-api/internal/infrastructure/logging"
	"github.com/gin-gonic/gin"
	"gorm.io/gorm"
)

type CommentHandler struct {
	service *services.CommentService
	logger  *logging.Logger
}

func NewCommentHandler(service *services.CommentService, logger *logging.Logger) *CommentHandler {
	return &CommentHandler{
		service: service,
		logger:  logger,
	}
}

// CreateComment handles POST /api/comments
//
// @Summary      Create a comment
// @Tags         comments
// @Accept       json
// @Produce      json
// @Param        comment  body      dtos.CreateCommentRequest  true  "Comment payload"
// @Success      201      {object}  dtos.SuccessResponse
// @Failure      400      {object}  dtos.ErrorResponse
// @Failure      404      {object}  dtos.ErrorResponse
// @Failure      500      {object}  dtos.ErrorResponse
// @Router       /comments [post]
func (h *CommentHandler) CreateComment(c *gin.Context) {
	var req dtos.CreateCommentRequest

	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, dtos.ErrorResponse{
			Error: dtos.ErrorDetail{
				Code:    "VALIDATION_ERROR",
				Message: "Request validation failed",
				Details: parseValidationErrors(err),
			},
		})
		return
	}

	comment, err := h.service.CreateComment(req)
	if err != nil {
		if containsStr(err.Error(), "invalid UUID") {
			c.JSON(http.StatusBadRequest, dtos.ErrorResponse{
				Error: dtos.ErrorDetail{
					Code:    "INVALID_COMMENT_ID",
					Message: "Invalid UUID format",
				},
			})
			return
		}

		if err == gorm.ErrRecordNotFound {
			c.JSON(http.StatusNotFound, dtos.ErrorResponse{
				Error: dtos.ErrorDetail{
					Code:    "COMMENT_NOT_FOUND",
					Message: "Comment with specified ID does not exist",
				},
			})
			return
		}

		if containsStr(err.Error(), "comments are disabled for whitenest chapters") {
			c.JSON(http.StatusForbidden, dtos.ErrorResponse{
				Error: dtos.ErrorDetail{
					Code:    "WHITENEST_COMMENTS_DISABLED",
					Message: "Comments are disabled for Whitenest chapters",
				},
			})
			return
		}

		h.logger.Error("Failed to create comment", logging.F("error", err.Error()))
		c.JSON(http.StatusInternalServerError, dtos.ErrorResponse{
			Error: dtos.ErrorDetail{
				Code:    "INTERNAL_ERROR",
				Message: "Failed to create comment",
			},
		})
		return
	}

	h.logger.Info("Comment created successfully", logging.F("id", comment.ID))
	c.JSON(http.StatusCreated, dtos.SuccessResponse{Data: comment})
}

// ListComments handles GET /api/comments
//
// @Summary      List comments
// @Tags         comments
// @Produce      json
// @Param        postId     query     string  false  "Filter by post ID"
// @Param        author     query     string  false  "Filter by author"
// @Param        sortBy     query     string  false  "Sort field"
// @Param        sortOrder  query     string  false  "asc or desc"
// @Success      200        {object}  dtos.CommentListResponse
// @Failure      400        {object}  dtos.ErrorResponse
// @Failure      500        {object}  dtos.ErrorResponse
// @Router       /comments [get]
func (h *CommentHandler) ListComments(c *gin.Context) {
	h.logger.Debug("Listing comments", logging.F("postId", c.Query("postId")), logging.F("author", c.Query("author")), logging.F("sortBy", c.Query("sortBy")), logging.F("sortOrder", c.Query("sortOrder")))
	filters := models.CommentFilters{
		PostID:    c.Query("postId"),
		Author:    c.Query("author"),
		SortBy:    c.Query("sortBy"),
		SortOrder: c.Query("sortOrder"),
	}
	comments, err := h.service.GetComments(filters)
	if err != nil {
		if containsStr(err.Error(), "invalid input syntax for type uuid") {
			c.JSON(http.StatusBadRequest, dtos.ErrorResponse{
				Error: dtos.ErrorDetail{
					Code:    "INVALID_COMMENT_ID",
					Message: "Invalid UUID format",
				},
			})
			return
		}

		c.JSON(http.StatusInternalServerError, dtos.ErrorResponse{
			Error: dtos.ErrorDetail{
				Code:    "INTERNAL_ERROR",
				Message: "Failed to list comments",
			},
		})
		return
	}

	c.JSON(http.StatusOK, comments)
}

// DeleteComment handles DELETE /api/comments/:id
//
// @Summary      Delete a comment
// @Tags         comments
// @Param        id  path  string  true  "Comment UUID"
// @Success      200  {object}  dtos.SuccessResponse
// @Failure      400  {object}  dtos.ErrorResponse
// @Failure      404  {object}  dtos.ErrorResponse
// @Failure      500  {object}  dtos.ErrorResponse
// @Router       /comments/{id} [delete]
func (h *CommentHandler) DeleteComment(c *gin.Context) {
	id := c.Param("id")
	if id == "" {
		c.JSON(http.StatusBadRequest, dtos.ErrorResponse{
			Error: dtos.ErrorDetail{
				Code:    "INVALID_COMMENT_ID",
				Message: "Invalid comment ID",
			},
		})
		return
	}

	if err := h.service.DeleteComment(id); err != nil {
		if err == gorm.ErrRecordNotFound {
			c.JSON(http.StatusNotFound, dtos.ErrorResponse{
				Error: dtos.ErrorDetail{
					Code:    "COMMENT_NOT_FOUND",
					Message: "Comment with specified ID does not exist",
				},
			})
			return
		}

		h.logger.Error("Failed to delete comment", logging.F("error", err.Error()))
		c.JSON(http.StatusInternalServerError, dtos.ErrorResponse{
			Error: dtos.ErrorDetail{
				Code:    "INTERNAL_ERROR",
				Message: "Failed to delete comment",
			},
		})
		return
	}

	h.logger.Info("Comment deleted successfully", logging.F("id", id))
	c.JSON(http.StatusOK, dtos.SuccessResponse{Data: id})
}